// typings/resume.d.ts
export interface Basics {
  /**
   * e.g. thomas@gmail.com
   */
  email?: string
  /**
   * URL (as per RFC 3986) to a image in JPEG or PNG format
   */
  image?: string
  /**
   * e.g. Web Developer
   */
  label?: string
  location?: Location
  name?: string
  /**
   * Phone numbers are stored as strings so use any format you like, e.g. 712-117-2923
   */
  phone?: string
  /**
   * Specify any number of social networks that you participate in
   */
  profiles?: Profile[]
  /**
   * Write a short 2-3 sentence biography about yourself
   */
  summary?: string
  /**
   * URL (as per RFC 3986) to your website, e.g. personal homepage
   */
  url?: string
}

export interface Location {
  /**
   * To add multiple address lines, use
   * . For example, 1234 Glücklichkeit Straße
   * Hinterhaus 5. Etage li.
   */
  address?: string
  city?: string
  /**
   * code as per ISO-3166-1 ALPHA-2, e.g. US, AU, IN
   */
  countryCode?: string
  postalCode?: string
  /**
   * The general region where you live. Can be a US state, or a province, for instance.
   */
  region?: string
}

export interface Profile {
  /**
   * e.g. Facebook or Twitter
   */
  network?: string
  /**
   * e.g. http://twitter.example.com/neutralthoughts
   */
  url?: string
  /**
   * e.g. neutralthoughts
   */
  username?: string
}